refactor 'RequestBuilder' so it can take arbitrary url params, and isn't specifically for referral requests
